backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-multipart==0.0.20
faiss-cpu==1.8.0.post1
fastembed==0.4.2
numpy==1.26.4
pypdf==5.1.0
requests==2.32.3
